refactor(vault-index): split tfidf.py + themes.py out of vault_index.py (#229)
Behavior-neutral extraction (Slice A of #53). Acyclic graph tfidf<-themes<-
vault_index; themes lazy-imports _connect to break the cycle; explicit named
re-export shim keeps vault_index.<symbol> resolving. _NEGATION_TERMS placed in
themes.py (corrects parent spec). Existing test_tfidf/themes/surprise pass
unmodified; adds test_module_split_parity.py.
