+++ hooks/vault_index.py
@@ -20,6 +20,24 @@ from collections import defaultdict
 from datetime import date
 from pathlib import Path
 
+# --- #229 Slice A: TF-IDF + theme primitives moved to sibling modules.
+# Re-exported here so `vault_index.<symbol>` keeps resolving for all callers,
+# and so extract_keywords() (below) can use the shared _STOPWORDS. ---
+from tfidf import (  # noqa: F401  (re-export shim — callers use vault_index.<symbol>)
+    _STOPWORDS,
+    _TOKEN_RE,
+    _tokenize_for_tfidf,
+    _compute_tfidf_vector,
+    _cosine_similarity,
+    _update_term_df,
+)
+from themes import (  # noqa: F401  (re-export shim)
+    assign_to_theme,
+    detect_surprise,
+    _NEGATION_TERMS,
+    _THEME_SIMILARITY_THRESHOLD,
+)
+
 # ---------------------------------------------------------------------------
 # Schema
 # ---------------------------------------------------------------------------
@@ -83,21 +101,6 @@ CREATE TABLE IF NOT EXISTS term_df (
     df              INTEGER NOT NULL
 );
 """
-
-# ---------------------------------------------------------------------------
-# Stopwords for keyword extraction
-# ---------------------------------------------------------------------------
-
-_STOPWORDS = frozenset(
-    "a an the and or but in on at to for of is it this that was were be been "
-    "being have has had do does did will would shall should may might can could "
-    "not no nor so if then than too very are am was were with from by about "
-    "into through during before after above below between out off over under "
-    "again further once here there when where why how all each every both few "
-    "more most other some such only own same also just because as until while "
-    "up down its they them their what which who whom he she we you i my your "
-    "his her our me him us".split()
-)
 
 # ---------------------------------------------------------------------------
 # Path helpers
@@ -1564,134 +1567,6 @@ def search_vault(
 
 
 # ---------------------------------------------------------------------------
-# TF-IDF primitives (stdlib-first, numpy optional)
-# ---------------------------------------------------------------------------
-
-
-_TOKEN_RE = re.compile(r"[a-z0-9]+")
-
-
-def _tokenize_for_tfidf(text: str) -> list[str]:
-    """Tokenize text for TF-IDF: lowercase alphanumerics, drop stopwords + single chars.
-
-    Order is preserved (token occurrences count — duplicates are kept) so the
-    caller can compute TF by counting. Returns [] for empty or all-stopword input.
-
-    Single-character tokens (both letters like "a" and digits like "3" or "9"
-    from expressions such as "Python-3.9") are dropped. Single digits appear
-    in few documents initially, which inflates their IDF and lets them
-    outrank real semantic terms in the sparse top-k — so letter and digit
-    noise are both removed uniformly.
-    """
-    if not text:
-        return []
-    lowered = text.lower()
-    return [
-        t for t in _TOKEN_RE.findall(lowered)
-        if len(t) > 1 and t not in _STOPWORDS
-    ]
-
-
-def _compute_tfidf_vector(
-    tokens: list[str],
-    term_df: dict[str, int],
-    total_docs: int,
-    top_k: int = 50,
-) -> dict[str, float]:
-    """Compute a sparse TF×IDF vector keeping the top_k heaviest terms.
-
-    tokens:       output of _tokenize_for_tfidf() for the document
-    term_df:      {term: document_frequency} for the corpus
-    total_docs:   total indexed documents (including this one if it is new;
-                  callers using incremental updates should increment total_docs
-                  BEFORE calling this function for a newly-inserted note)
-    top_k:        maximum number of terms to retain in the sparse vector
-
-    Returns {} when tokens is empty. Uses smoothed IDF
-    (1 + ln((N + 1) / (df + 1))) so new / rare terms never produce 0 or NaN.
-    """
-    if not tokens:
-        return {}
-
-    tf: dict[str, int] = defaultdict(int)
-    for t in tokens:
-        tf[t] += 1
-
-    weights: dict[str, float] = {}
-    n_plus_one = total_docs + 1
-    for term, count in tf.items():
-        df = term_df.get(term, 0)
-        idf = 1.0 + math.log(n_plus_one / (df + 1))
-        weights[term] = count * idf
-
-    top = sorted(weights.items(), key=lambda kv: (-kv[1], kv[0]))[:top_k]
-    return dict(top)
-
-
-def _cosine_similarity(v1: dict[str, float], v2: dict[str, float]) -> float:
-    """Cosine similarity between two sparse dict vectors. Returns 0.0 on empty input.
-
-    Iterates the smaller dict to compute the dot product, which keeps the
-    inner loop bounded even when one vector is much larger than the other.
-    """
-    if not v1 or not v2:
-        return 0.0
-
-    if len(v1) > len(v2):
-        v1, v2 = v2, v1
-
-    dot = 0.0
-    for term, w in v1.items():
-        other = v2.get(term)
-        if other is not None:
-            dot += w * other
-
-    if dot == 0.0:
-        return 0.0
-
-    norm1 = math.sqrt(sum(w * w for w in v1.values()))
-    norm2 = math.sqrt(sum(w * w for w in v2.values()))
-    if norm1 == 0.0 or norm2 == 0.0:
-        return 0.0
-    return dot / (norm1 * norm2)
-
-
-def _update_term_df(
-    conn: sqlite3.Connection,
-    old_terms: set[str],
-    new_terms: set[str],
-) -> None:
-    """Adjust document-frequency counts for a note whose terms changed.
-
-    Compares the two sets and applies a +1 / -1 per term via executemany.
-    Terms whose df falls to zero are deleted so the IDF denominator stays
-    clean. The caller is responsible for committing the transaction — this
-    helper writes through ``conn`` but does not commit, so it can be
-    batched atomically with a note upsert or delete.
-    """
-    removed = old_terms - new_terms
-    added = new_terms - old_terms
-    if not removed and not added:
-        return
-
-    cur = conn.cursor()
-
-    if added:
-        cur.executemany(
-            "INSERT INTO term_df (term, df) VALUES (?, 1) "
-            "ON CONFLICT(term) DO UPDATE SET df = df + 1",
-            [(t,) for t in added],
-        )
-
-    if removed:
-        cur.executemany(
-            "UPDATE term_df SET df = df - 1 WHERE term = ?",
-            [(t,) for t in removed],
-        )
-        cur.execute("DELETE FROM term_df WHERE df <= 0")
-
-
-# ---------------------------------------------------------------------------
 # Keyword extraction
 # ---------------------------------------------------------------------------
 
@@ -1852,217 +1727,3 @@ def query_related_notes(
     finally:
         conn.close()
 
-
-# ---------------------------------------------------------------------------
-# Theme assignment (incremental)
-# ---------------------------------------------------------------------------
-
-
-_THEME_SIMILARITY_THRESHOLD = 0.3
-
-
-def assign_to_theme(
-    db_path: str,
-    note_path: str,
-    project: str | None = None,
-    similarity_threshold: float = _THEME_SIMILARITY_THRESHOLD,
-) -> dict | None:
-    """Incrementally assign a summarized note to its nearest theme (if any).
-
-    Reads the note's tfidf_vector and compares against every project-scoped
-    theme centroid plus any cross-project themes (project IS NULL). If the
-    best similarity exceeds ``similarity_threshold``, adds a theme_members
-    row and updates the theme's centroid via running average.
-
-    Returns {"theme_id": int, "similarity": float} on assignment, or None
-    if no theme was close enough (or the note has no vector).
-    """
-    if not os.path.isfile(db_path):
-        return None
-
-    try:
-        conn = _connect(db_path)
-    except sqlite3.Error as exc:
-        print(f"[vault-index] assign_to_theme could not connect: {exc}",
-              file=sys.stderr)
-        return None
-
-    try:
-        # Take the write lock BEFORE reading candidates so the
-        # read-modify-write cycle is atomic against concurrent writers.
-        # Other callers block for up to the connection's timeout (5s in
-        # _connect()) before sqlite3 raises OperationalError.
-        conn.execute("BEGIN IMMEDIATE")
-        try:
-            row = conn.execute(
-                "SELECT tfidf_vector FROM notes WHERE path = ?", (note_path,)
-            ).fetchone()
-            if not row or not row["tfidf_vector"]:
-                conn.commit()
-                return None
-            try:
-                note_vec = json.loads(row["tfidf_vector"])
-            except json.JSONDecodeError:
-                conn.commit()
-                return None
-            if not note_vec:
-                conn.commit()
-                return None
-
-            # When project is None (global/unscoped recall), "project = ?"
-            # binds to NULL and matches no rows because `NULL = NULL` is
-            # false in SQL — only project-scoped themes with project IS
-            # NULL would match. Branch the query so unscoped callers see
-            # every theme, and scoped callers see their own + cross-project.
-            if project is None:
-                candidates = conn.execute(
-                    "SELECT id, centroid, note_count FROM themes"
-                ).fetchall()
-            else:
-                candidates = conn.execute(
-                    "SELECT id, centroid, note_count FROM themes "
-                    "WHERE project = ? OR project IS NULL",
-                    (project,),
-                ).fetchall()
-
-            best: tuple[float, int, dict, int] | None = None
-            for cand in candidates:
-                if not cand["centroid"]:
-                    continue
-                try:
-                    centroid = json.loads(cand["centroid"])
-                except json.JSONDecodeError:
-                    continue
-                sim = _cosine_similarity(note_vec, centroid)
-                if best is None or sim > best[0]:
-                    best = (sim, cand["id"], centroid, cand["note_count"])
-
-            if best is None or best[0] < similarity_threshold:
-                conn.commit()
-                return None
-
-            sim, theme_id, centroid, count = best
-            today = date.today().isoformat()
-
-            # Detect reassignment: if the note is already a member, the
-            # centroid already reflects its contribution — updating count
-            # and re-averaging would double-count it and drift the centroid.
-            already_member = conn.execute(
-                "SELECT 1 FROM theme_members "
-                "WHERE theme_id = ? AND note_path = ?",
-                (theme_id, note_path),
-            ).fetchone() is not None
-
-            if not already_member:
-                new_centroid: dict[str, float] = {}
-                all_terms = set(centroid) | set(note_vec)
-                for term in all_terms:
-                    c_val = centroid.get(term, 0.0)
-                    v_val = note_vec.get(term, 0.0)
-                    new_centroid[term] = (c_val * count + v_val) / (count + 1)
-
-                conn.execute(
-                    "UPDATE themes "
-                    "SET centroid = ?, note_count = ?, updated_date = ? "
-                    "WHERE id = ?",
-                    (json.dumps(new_centroid, separators=(",", ":")),
-                     count + 1, today, theme_id),
-                )
-            else:
-                # Reassignment: bump updated_date but leave count/centroid
-                # alone. The member's similarity is refreshed below.
-                conn.execute(
-                    "UPDATE themes SET updated_date = ? WHERE id = ?",
-                    (today, theme_id),
-                )
-
-            # Preserve surprise + added_date on reassignment — only
-            # similarity is refreshed from the latest cosine computation.
-            conn.execute(
-                "INSERT INTO theme_members "
-                "(theme_id, note_path, similarity, surprise, added_date) "
-                "VALUES (?, ?, ?, 0.0, ?) "
-                "ON CONFLICT(theme_id, note_path) DO UPDATE SET "
-                "similarity = excluded.similarity",
-                (theme_id, note_path, sim, today),
-            )
-            conn.commit()
-            return {"theme_id": theme_id, "similarity": sim}
-        except Exception:
-            conn.rollback()
-            raise
-    finally:
-        conn.close()
-
-
-_NEGATION_TERMS = frozenset({
-    # Simple negation / failure words
-    "not", "never", "failed", "broken", "wrong", "mistake",
-    "avoid", "no", "cannot",
-    # Contractions — bare forms only. detect_surprise() strips apostrophes
-    # from note_text before tokenizing, so "don't"/"won't"/"isn\u2019t"
-    # in source text collapse to "dont"/"wont"/"isnt" and match here.
-    # Apostrophed variants (e.g. "don't") would never match post-
-    # tokenization and are intentionally omitted.
-    "dont", "cant", "wont", "isnt", "arent", "wasnt", "werent",
-    "didnt", "doesnt", "couldnt", "shouldnt", "wouldnt",
-    "hasnt", "havent", "hadnt",
-})
-
-
-def detect_surprise(
-    note_text: str,
-    note_vec: dict[str, float],
-    theme_centroid: dict[str, float],
-    window: int = 8,
-    top_shared: int = 10,
-) -> float:
-    """Heuristic Free-Energy surprise score for a note vs. its theme centroid.
-
-    Returns the fraction of the top_shared shared TF-IDF terms that appear
-    within ``window`` tokens of a negation word in ``note_text``. Clamped
-    to [0.0, 1.0]. Zero on missing overlap or empty input.
-    """
-    if not note_text or not note_vec or not theme_centroid:
-        return 0.0
-
-    shared = [
-        (t, min(note_vec[t], theme_centroid[t]))
-        for t in set(note_vec) & set(theme_centroid)
-    ]
-    if not shared:
-        return 0.0
-    shared.sort(key=lambda kv: (-kv[1], kv[0]))
-    shared_terms = [t for t, _ in shared[:top_shared]]
-
-    # Strip apostrophes (straight + smart) before tokenizing so contractions
-    # like "don't" / "can't" collapse to "dont"/"cant" and match the
-    # negation set. Without this, _TOKEN_RE = [a-z0-9]+ would emit
-    # "don", "t" and the negation check would silently miss every
-    # apostrophed negation in the source text.
-    normalized = (
-        note_text.lower()
-        .replace("\u2019", "")  # right single quote (smart apostrophe)
-        .replace("\u2018", "")  # left single quote
-        .replace("'", "")       # straight apostrophe
-    )
-    tokens = _TOKEN_RE.findall(normalized)
-    if not tokens:
-        return 0.0
-
-    negation_positions = [
-        i for i, t in enumerate(tokens) if t in _NEGATION_TERMS
-    ]
-    if not negation_positions:
-        return 0.0
-
-    hits = 0
-    for term in shared_terms:
-        term_positions = [i for i, t in enumerate(tokens) if t == term]
-        for p in term_positions:
-            if any(abs(p - n) <= window for n in negation_positions):
-                hits += 1
-                break
-
-    score = hits / len(shared_terms)
-    return max(0.0, min(1.0, score))
